requirements.txt
# Web scraping
beautifulsoup4>=4.14.0
lxml>=4.9.0
requests>=2.32.0
selenium>=4.36.0
webdriver-manager>=4.0.0

# Database and storage
minio>=7.2.0
psycopg2-binary>=2.9.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning and preprocessing
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

